requirements.txt
face_recognition==1.3.0
matplotlib==3.5.1
numpy==1.21.4
opencv_python==4.5.4.60
pandas==1.3.5
Pillow==8.4.0
streamlit==1.3.0
